Remove body from invoices
The Zuora API only allows fetching the body field on a single invoice at a time
which breaks fetching collections of invoices.

This removes the body field from regular queries, allowing collections to be
fetched without errors. In order to allow the PDF of the invoice (which is
normally stored in the body) to be accessed without having to write a specific
query, this also adds a `pdf` method to invoices.